requirements.txt
# LED Matrix Project - Python Dependencies
# Install: pip install -r requirements.txt
#
# Tested platforms:
#   - Raspberry Pi OS (Bookworm, Python 3.11, ARM64)
#   - Windows 10/11 (Python 3.9-3.12, x64)
#   - macOS (Python 3.9-3.12, Apple Silicon / Intel)
#
# Note: On Raspberry Pi, opencv-python-headless is used instead of
# opencv-python to avoid Qt/GTK GUI dependencies. Both provide
# identical cv2 API for frame processing.

# Image processing
Pillow>=10.0.0

# Video processing (for YouTube stream feature)
# On Raspberry Pi: if pip install fails, use system package instead:
#   sudo apt-get install python3-opencv
opencv-python-headless>=4.8.0
numpy>=1.24.0
# Install yt-dlp from GitHub source (gets extraction fixes hours before PyPI)
yt-dlp @ https://github.com/yt-dlp/yt-dlp/archive/refs/tags/2025.03.31.tar.gz

# HTTP requests (for Bitcoin price, connectivity checks)
requests>=2.31.0

# Development: LED matrix simulator (not needed on Raspberry Pi)
# Install on dev machines: pip install pygame
pygame>=2.5.0; sys_platform == "win32" or sys_platform == "darwin" or sys_platform == "linux"

# Web control panel
Flask>=3.0.0
flask-sock>=0.7.0

# QR code generation
qrcode>=7.4

# System resource monitoring (CPU, RAM, disk, temperature)
psutil>=5.9.0

# Testing: moved to requirements-dev.txt
# Install with: pip install -r requirements-dev.txt
